offer book view checks for invalid signer or signature
merge OfferFilter into OfferFilterService
