+++ core/src/main/resources/i18n/displayStrings.properties
@@ -415,6 +415,9 @@ offerbook.warning.requireUpdateToNewVersion=Your version of Haveno is not compat
 offerbook.warning.offerWasAlreadyUsedInTrade=You cannot take this offer because you already took it earlier. \
   It could be that your previous take-offer attempt resulted in a failed trade.
 
+offerbook.warning.arbitratorNotValidated=This offer cannot be taken because the arbitrator is invalid
+offerbook.warning.signatureNotValidated=This offer cannot be taken because the arbitrator's signature is invalid
+
 offerbook.info.sellAtMarketPrice=You will sell at market price (updated every minute).
 offerbook.info.buyAtMarketPrice=You will buy at market price (updated every minute).
 offerbook.info.sellBelowMarketPrice=You will get {0} less than the current market price (updated every minute).
